feat: add app icon, reset form after submission, and update build settings
This commit introduces the app icon to both Android and iOS platforms, adds the functionality to reset the sell item form after submission, and updates the asset compiler settings in the iOS project.
WARNING: Icons with alpha channel are not allowed in the Apple App Store.
Set "remove_alpha_ios: true" to remove it.

- Added `flutter_launcher_icons` dependency to `pubspec.yaml`.
- Added app icon image to `lib/assets/icons/Mozayed_app_icon.png`.
- Configured app icons for Android and iOS in `pubspec.yaml`.
- Updated `sell_screen.dart` to reset the form after successful submission.
- Modified `project.pbxproj` to set `ASSETCATALOG_COMPILER_GENERATE_SWIFT_ASSET_SYMBOL_EXTENSIONS` to `AppIcon`.
- Added new icon files to `android/app/src/main/res` and `ios/Runner/Assets.xcassets/AppIcon.appiconset/` .
- Deleted `ios/Runner/Assets.xcassets/AppIcon.appiconset/Contents.json` file and updated it's contents.
- Updated `pubspec.lock` dependencies.